+++ Presentation/Feature/PCWebView/Project.swift
@@ -1,0 +1,14 @@
+//
+// Project.swift
+// PCWebView
+//
+// Created by summercat on 2025/02/12.
+//
+
+import ProjectDescription
+import ProjectDescriptionHelpers
+
+let project = Project.dynamicFramework(
+  name: Modules.Presentation.PCWebView.rawValue,
+  dependencies: []
+)
