+++ Tuist/ProjectDescriptionHelpers/Modules.swift
@@ -73,9 +73,13 @@ public extension Modules {
 
 public extension Modules {
   enum Presentation: String {
+    // 공통
     case DesignSystem
     case Router
     case Coordinator
+    case PCWebView
+    
+    // 피처
     case Login
     case SignUp
     case Home
